+++ CHANGELOG.md
@@ -188,3 +188,18 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### Added
 -test player: casos límite de fichas sacadas y en barra
+
+#### [0.3.6] - 2024-09-27
+
+### Added
+-test board:tests de inicialización y movimientos
+
+#### [0.3.7] - 2024-09-27
+
+### Added
+-test board: mover ficha, capturas, barra y reingresos
+
+#### [0.3.8] - 2024-09-27
+
+### Added
+-test board: agregar casos para bear off (puede_sacar_fichas y sacar_ficha_del_tablero)